redefinition of the long SUMOTime to avoid using cstdint (which requires c++11), refs #1728
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@18239 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSNet.h
@@ -647,7 +647,7 @@ protected:
     long mySimBeginMillis;
 
     /// @brief The overall number of vehicle movements
-    int_fast64_t myVehiclesMoved;
+    long long int myVehiclesMoved;
     //}
