+++ sumo/src/netimport/NIImporter_VISUM.h
@@ -502,16 +502,16 @@ private:
 
 
     /// @brief A map of point ids to positions
-    std::map<int_fast64_t, Position> myPoints;
+    std::map<long long int, Position> myPoints;
 
     /// @brief A map of edge (not road, but "edge" in this case) ids to from/to-points
-    std::map<int_fast64_t, std::pair<int_fast64_t, int_fast64_t> > myEdges;
+    std::map<long long int, std::pair<long long int, long long int> > myEdges;
 
     /// @brief A map from district shape definition name to the district
-    std::map<int_fast64_t, NBDistrict*> myShapeDistrictMap;
+    std::map<long long int, NBDistrict*> myShapeDistrictMap;
 
     /// @brief A map from area parts to area ids
-    std::map<int_fast64_t, std::vector<int_fast64_t> > mySubPartsAreas;
+    std::map<long long int, std::vector<long long int> > mySubPartsAreas;
 
     /// @brief A temporary storage for district shapes as they are filled incrementally
     std::map<NBDistrict*, PositionVector> myDistrictShapes;
