+++ sumo/src/utils/xml/SUMOSAXAttributesImpl_Binary.h
@@ -133,7 +133,7 @@ public:
      * @exception EmptyData If the attribute is not known or the attribute value is an empty string
      * @exception NumberFormatException If the attribute value can not be parsed to an int
      */
-    int_fast64_t getLong(int id) const;
+    long long int getLong(int id) const;
 
 
     /**
